Build-rust-deb-action initial commit
Initial commit for a new github action to build Rust projects for Debain
based Linux distributions. This should leverage dedicated Rust
environment and make use of native Debian gcc and glibc as well as other
native libraries, if required. Cross compilation to any architecture
should be possible. Custom additional build tools for rust and Debian
can be installed. To preserve a clean environment, everythin is run
inside a docker container. As a similar project exists for native
Debian crossbuilds, this project was established as a fork of one that
looked promising and was heavily used here as well (shout out to
https://github.com/jtdor/build-deb-action)